download rest api endpoints, general code refactor
